pyproject.toml
[project]
name = "travel-agent-app"
version = "0.1.0"
description = "Multi-agent event-driven system to mitigate LLM knowledge cutoff in dynamic urban contexts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "opentelemetry-api>=1.26.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
]
nats = [
  "nats-py>=2.7.2",
]
otel = [
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
